requirements.txt
amqp==2.2.2
APScheduler==3.3.1
argon2-cffi==16.3.0
bcrypt==3.1.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-allauth==0.33.0
django-bootstrap3==9.0.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.6.1
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
maxminddb==1.3.0
oauthlib==2.0.4
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.3.1
pycparser==2.18
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
static3==0.7.0
tzlocal==1.4
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
